% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_nodes}
\alias{layout_nodes}
\title{Extract the node coordinates from a layout}
\usage{
layout_nodes(layout)
}
\arguments{
\item{layout}{A network layout.}
}
\value{
A data frame with \code{name}, \code{x}, and \code{y}.
}
\description{
Extract the node coordinates from a layout
}
