% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{load_network_layout}
\alias{load_network_layout}
\title{Load and validate cached network coordinates}
\usage{
load_network_layout(path, network)
}
\arguments{
\item{path}{Layout CSV path.}

\item{network}{A \code{hybs_network} object.}
}
\value{
A \code{hybs_network_layout} data frame.
}
\description{
Load and validate cached network coordinates
}
