% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_edges}
\alias{network_edges}
\title{Extract network edge data}
\usage{
network_edges(network)
}
\arguments{
\item{network}{A \code{hybs_network} object.}
}
\value{
An edge data frame.
}
\description{
Extract network edge data
}
