% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_nodes}
\alias{network_nodes}
\title{Extract network node data}
\usage{
network_nodes(network)
}
\arguments{
\item{network}{A \code{hybs_network} object.}
}
\value{
A node data frame.
}
\description{
Extract network node data
}
