% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_local_network}
\alias{plot_local_network}
\title{Plot a focal induced network}
\usage{
plot_local_network(
  network,
  layout = NULL,
  style = theme_fig3_current(),
  title = NULL,
  seed = 42L
)
}
\arguments{
\item{network}{A local \code{hybs_network} returned by \code{\link[=extract_local_network]{extract_local_network()}}.}

\item{layout}{Cached or newly calculated layout.}

\item{style}{A \code{hybs_network_style} object.}

\item{title}{Optional title.}

\item{seed}{Seed for layout and labels.}
}
\value{
A ggplot object.
}
\description{
Plot a focal induced network
}
