% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{read_hybs_csv}
\alias{read_hybs_csv}
\title{Read and validate a CSV table}
\usage{
read_hybs_csv(path, required = character())
}
\arguments{
\item{path}{Path to a CSV file.}

\item{required}{Required column names.}
}
\value{
A data frame.
}
\description{
Read and validate a CSV table
}
