% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{read_network_config}
\alias{read_network_config}
\title{Read an editable YAML configuration}
\usage{
read_network_config(path)
}
\arguments{
\item{path}{YAML path.}
}
\value{
A named list.
}
\description{
Read an editable YAML configuration
}
