% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{theme_fig3_colorblind}
\alias{theme_fig3_colorblind}
\title{Colour-blind-friendly Figure 3 network style}
\usage{
theme_fig3_colorblind()
}
\value{
A \code{hybs_network_style} object.
}
\description{
Colour-blind-friendly Figure 3 network style
}
