% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{validate_network_layout}
\alias{validate_network_layout}
\title{Validate layout coordinates against network topology}
\usage{
validate_network_layout(layout, network)
}
\arguments{
\item{layout}{Layout data frame.}

\item{network}{A \code{hybs_network} object.}
}
\value{
The layout invisibly.
}
\description{
Validate layout coordinates against network topology
}
